[hotfix][tests] Merge DefaultSchedulerBuilder and AdaptiveBatchSchedulerBuilder
The extending AdaptiveBatchSchedulerBuilder is in a good shape, which requires hack logics to create AdaptiveBatchScheduler. Merging them can make it simpler to create different schedulers.
